i9100 - Add Overlay to activate TouchKey Delay Menu Option.
Commit to framework & Settings implements TouchKey Timeout Delay settings but as some devices do not have backlit TouchKeys this Setting must be enabled through the overlay to set the option to true.

Any other device that would be compatible with this would just need to apply the same additions to their device.

Change-Id: I3a77554370ac57b7ed943d8668c766da0ffb1cfa